+++ tests/test_frame.py
@@ -11,6 +11,7 @@ from pyppeteer.errors import ElementHandleError, TimeoutError
 
 from .base import BaseTestCase
 from .frame_utils import attachFrame, detachFrame, dumpFrames, navigateFrame
+from .utils import waitEvent
 
 addElement = 'tag=>document.body.appendChild(document.createElement(tag))'
 
@@ -561,6 +562,15 @@ http://localhost:{port}/static/nested-frames.html
         await detachFrame(self.page, 'frame1')
         self.assertEqual(len(detachedFrames), 1)
         self.assertTrue(detachedFrames[0].isDetached())
+
+    @sync
+    async def test_anchor_url(self):
+        await self.page.goto(self.url + 'empty')
+        await asyncio.wait([
+            self.page.goto(self.url + 'empty#foo'),
+            waitEvent(self.page, 'framenavigated'),
+        ])
+        self.assertEqual(self.page.url, self.url+'empty#foo')
 
     @sync
     async def test_frame_cross_process(self):